requirements.txt
Django==1.5
Fabric==1.6.0
South==0.7.6
Pillow==2.1.0
django-localflavor==1.0
easy-thumbnails==1.4
paramiko==1.10.0
psycopg2==2.4.6
pycrypto==2.6
wsgiref==0.1.2

# Django Tiny-Cart
# git+https://github.com/trilan/django-tinycart.git#egg=django-tinycart
# Django Custom Auth
git+https://github.com/endthestart/django-custom-auth#egg=django-custom-auth
